% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_embeddings.R
\name{dynamic_embeddings}
\alias{dynamic_embeddings}
\title{Dynamic Embeddings Class}
\usage{
dynamic_embeddings(
  fcm = array(),
  context_embeddings = array(),
  word_embeddings = NULL,
  control = dynamic_embeddings_control(),
  train_method = NULL
)
}
\arguments{
\item{fcm}{a \link[quanteda:fcm]{Quanteda fcm} or similar 2D matrix-like or 3D
array-like object}

\item{context_embeddings}{a matrix or array with the same number of rows as \code{fcm}}

\item{word_embeddings}{a matrix or array with the same number of rows as \code{fcm}}

\item{control}{a list of control parameters created by \code{dynamic_embeddings_control()}}

\item{train_method}{the method used to train the embeddings}
}
\description{
A dynamic_embeddings object is a list containing embeddings and/or
a specification of the embedding method. This allows methods such as
\code{predict()}, which may in some cases compute embeddings dynamically to
avoid maintaining too bulky a model object. dynamic_embeddings objects are
generally produced as the output of a \code{train_} function (e.g. \code{train_svd()})
rather than being constructed directly.
}
