% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_embeddings_control.R
\name{dynamic_embeddings_control}
\alias{dynamic_embeddings_control}
\title{Control Parameters for Dynamic Embeddings Objects}
\usage{
dynamic_embeddings_control(
  control = NULL,
  train_method = NULL,
  drop_zeros = NULL,
  mlm_engine = NULL
)
}
\arguments{
\item{control}{(optional) an existing list of control parameters to modify}

\item{train_method}{the method used to train the embeddings; one of \code{"svd"},
\code{"mds"}, \code{"sgns"}, \code{"glove"}, \code{"lm"}, or \code{"mlm"}.}

\item{drop_zeros}{when \code{train_method = "lm"} or \code{"mlm"} should zeros be
dropped from fcm while fitting the model?}

\item{mlm_engine}{engine for fitting random effects models when
\code{train_method = "mlm"}. Either \code{"mgcv"} (the default), \code{"lme4"}, or \code{"jglmm"}
(requires Julia and the \code{jglmm} package).}
}
\description{
Control Parameters for Dynamic Embeddings Objects
}
