% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm_match.R
\name{fcm_match}
\alias{fcm_match}
\title{Match the feature set of an feature co-occurence matrix to given feature names}
\usage{
fcm_match(fcm, features, ...)
}
\arguments{
\item{fcm}{a \link[quanteda:fcm]{Quanteda fcm} or matrix-like object with row and column names}

\item{features}{a character vector of feature names}

\item{...}{additional arguments (not used)}
}
\value{
A matrix-like object with row and column names matching the given feature names
}
\description{
Match the feature set of a \link[quanteda:fcm]{Quanteda fcm} (or similar matrix-like
object) to a specified vector of feature names. For existing features in x for
which there is an exact match for an element of features, these will be included.
Any features in x not features will be discarded, and any feature names specified
in features but not found in x will be added with all zero counts.
}
