% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcm_reduce.R
\name{fcm_reduce}
\alias{fcm_reduce}
\title{Reduce a 3-D array to a 2-D feature co-occurrence matrix}
\usage{
fcm_reduce(fcm, f = `+`)
}
\arguments{
\item{fcm}{a \link[quanteda:fcm]{Quanteda fcm} or similar 2D matrix-like or 3D array-like object}

\item{f}{a binary aggregation function to be passed to \code{base::Reduce()}}
}
\description{
Reduce a N × N × M (sparse) array to an N × N matrix by summing (or some
other function) along the third dimension. Thus if \code{fcm} represents a stack
of feature co-occurrence matrices (FCMs), \code{fcm_reduce()} will compute the
aggregate FCM.
}
