% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_embeddings-predict.R
\name{predict.dynamic_embeddings}
\alias{predict.dynamic_embeddings}
\title{Retrieve Word Embeddings}
\usage{
\method{predict}{dynamic_embeddings}(object, newdata, drop = TRUE, .keep_missing = FALSE)
}
\arguments{
\item{object}{a \code{dynamic_embeddings} object}

\item{newdata}{a character vector of words}

\item{drop}{logical. If \code{TRUE} (the default) and the result is one-dimensional
(e.g. a single row), the output will be a (named) vector.}

\item{.keep_missing}{logical. What should be done about items in \code{newdata}
that are not present in the embeddings object? If \code{FALSE} (the default),
they will be ignored. If \code{TRUE}, they will be returned as \code{NA}.}
}
\description{
Retrieve embeddings of particular words from a trained model. For the
\code{\link[=train_lm]{"lm"}} and \code{"mlm"} training method,s embeddings may be computed
dynamically when requested.
}
\details{
Duplicated items in newdata will result in duplicated rows in the output.
If an item in newdata matches multiple rows in object, the last one will be
returned.
}
