% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_lm.R
\name{train_lm}
\alias{train_lm}
\title{Train embeddings with linear regression}
\usage{
train_lm(
  fcm,
  fcm_weight = NULL,
  n_dims = 100,
  context_embeddings = train_svd_context,
  ...,
  dynamic = TRUE,
  control = dynamic_embeddings_control()
)
}
\arguments{
\item{fcm}{a \link[quanteda:fcm]{Quanteda fcm} or similar 2D matrix-like or 3D
array-like object}

\item{fcm_weight}{(optional) an array-like object of the same dimensionality
as \code{fcm}, giving weights for the regression.}

\item{n_dims}{integer. Dimensionality of embeddings. Ignored if
\code{context_embeddings} is a matrix or array.}

\item{context_embeddings}{either a matrix or array with the same number of
rows as \code{fcm}, or a function that takes an FCM and outputs such a matrix or
array.}

\item{...}{additional parameters to be passed to \code{context_embeddings} function}

\item{dynamic}{logical. \code{TRUE} (the default) produces a \code{dynamic_embeddings}
object from which word embeddings can be computed dynamically as needed.
\code{FALSE} computes all word embeddings upfront.}

\item{control}{a list of control parameters created by \code{dynamic_embeddings_control()}}
}
\description{
Given a set of context embeddings and a feature co-occurence matrix (FCM),
an embedding for a given word can be trained using multiple linear regression
with the context embedding dimensions as predictors and the corresponding row
of the FCM as the dependent variable. This method enables the quantification
of uncertainty around the resulting embeddings (Vallebueno et al., 2024).
}
\details{
For GloVe-V (Vallebueno et al., 2024), \code{fcm} should be a matrix of log
co-occurrence counts with pretrained word and context biases subtracted out,
\code{fcm_weight} should be a matrix of transformed co-occurrence counts such that
low frequencies are devalued, and \code{context_embeddings} should be pretrained
glove context embeddings.
}
\references{
Vallebueno, A., Handan-Nader, C., Manning, C. D., & Ho, D. E. (2024). Statistical Uncertainty in Word Embeddings: GloVe-V (arXiv:2406.12165). arXiv. https://doi.org/10.48550/arXiv.2406.12165
}
