% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_mlm.R
\name{train_mlm}
\alias{train_mlm}
\title{Train embeddings with multilevel linear regression}
\usage{
train_mlm(
  fcm,
  fcm_weight = NULL,
  n_dims = 100,
  context_embeddings = train_svd_context,
  ...,
  dynamic = TRUE,
  control = dynamic_embeddings_control()
)
}
\arguments{
\item{fcm}{a 3D array-like object}

\item{fcm_weight}{(optional) an array-like object of the same dimensionality
as \code{fcm}, giving weights for the regression.}

\item{n_dims}{integer. Dimensionality of embeddings. Ignored if
\code{context_embeddings} is a matrix or array.}

\item{context_embeddings}{either a matrix or array with the same number of
rows as \code{fcm}, or a function that takes an FCM and outputs such a matrix or
array.}

\item{...}{additional parameters to be passed to \code{context_embeddings} function}

\item{dynamic}{logical. \code{TRUE} (the default) produces a \code{dynamic_embeddings}
object from which word embeddings can be computed dynamically as needed.
\code{FALSE} computes all word embeddings upfront.}

\item{control}{a list of control parameters created by \code{dynamic_embeddings_control()}}
}
\description{
This training paradigm follows the same principle as \code{train_lm()}, but
operates on a stacked (3D) feature co-occurrence matrix representing
co-occurrence patterns in a set of communities or individuals. Each
community or individual has distinct word embeddings, but these are assumed
to come from a single underlying normal distribution and therefore
information can be efficiently shared between communities or individuals
using partial pooling.
}
