% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_svd.R
\name{train_svd}
\alias{train_svd}
\alias{train_svd_context}
\title{Train embeddings with SVD}
\usage{
train_svd(
  fcm,
  n_dims = 100,
  eig = 0,
  row_weights = NULL,
  col_weights = NULL,
  output = "word_embeddings",
  opts = list()
)

train_svd_context(fcm, n_dims = 100, eig = 0)
}
\arguments{
\item{fcm}{a \link[quanteda:fcm]{Quanteda fcm} or similar 2D matrix-like or 3D
array-like object}

\item{n_dims}{integer. Dimensionality of embeddings.}

\item{eig}{numeric. Exponent for weighting the eigenvalue matrix, as per
Caron (2001). The default, \code{0}, corresponds to ignoring the eigenvalues
entirely.}

\item{row_weights}{numeric vector. Optional weights for the rows of \code{fcm}.}

\item{col_weights}{numeric vector. Optional weights for the columns of \code{fcm}.}

\item{output}{character. The default, \code{"word_embeddings"} includes only word
embeddings in the output. \code{"context_embeddings"} includes only context
embeddings. \code{"all"} includes both word and context embeddings.}

\item{opts}{control parameters passed to \code{\link[RSpectra:svds]{RSpectra::svds()}}.}
}
\description{
\code{train_svd(fcm)} decomposes the input \code{fcm} to produce word and/or context
embeddings. \code{train_svd_context(fcm)} is equivalent to
\code{train_svd(fcm, output = "context_embeddings")$context_embeddings}.
}
