% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordembeddings-package.R
\docType{package}
\name{wordembeddings-package}
\alias{wordembeddings}
\alias{wordembeddings-package}
\title{wordembeddings: Create customized word embeddings using a grammar of model (hyper)parameters}
\description{
The various algorithms for training word embeddings (e.g. word2vec, GloVe, fastText, PPMI SVD) share common underlying principles, but these principles are often obscured by the choices made in particular implementations. 'wordembeddings' makes the assumptions underlying these algorithms explicit with a comprehensive grammar of model parameters and hyperparameters. The user can train embeddings using any of the traditional frameworks, or mix and match elements to fit a particular application.
}
\author{
\strong{Maintainer}: Louis Teitelbaum \email{louist@post.bgu.ac.il} (\href{https://orcid.org/0009-0001-9347-0145}{ORCID}) [copyright holder]

}
\keyword{internal}
